pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "piholecombinelist"
version = "1.8.0"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "customtkinter>=5.2.0",
    "Pillow>=9.0.0",
]

[project.scripts]
phlist = "piholecombinelist.gui:main"
phlist-desktop = "piholecombinelist._install_desktop:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
piholecombinelist = ["assets/*"]
